requirements.txt
fastapi
uvicorn
python-dotenv
httpx
slowapi
google-generativeai
duckduckgo-search